Find note definition by class

+++ src/java/co/com/soinsoftware/altablero/entity/NoteDefinitionBO.java
@@ -5,7 +5,6 @@
  */
 package co.com.soinsoftware.altablero.entity;
 
-import java.math.BigDecimal;
 import java.util.Set;
 
 /**
@@ -19,7 +18,7 @@ public class NoteDefinitionBO extends AbstractBO {
 
     private String description;
 
-    private BigDecimal value;
+    private Integer value;
 
     private PeriodBO period;
 
@@ -37,11 +36,11 @@ public class NoteDefinitionBO extends AbstractBO {
         this.description = description;
     }
 
-    public BigDecimal getValue() {
+    public Integer getValue() {
         return value;
     }
 
-    public void setValue(BigDecimal value) {
+    public void setValue(Integer value) {
         this.value = value;
     }
